fixed javadoc
git-svn-id: file:///home/hoijui/src/IsaViz/tmp2/zvtm.svn/code@3571 74e165f3-c697-40a6-91af-552c8936d7fe

+++ zvtm-pdf/trunk/src/main/java/fr/inria/zvtm/glyphs/ZPDFPage.java
@@ -94,12 +94,12 @@ public abstract class ZPDFPage extends ClosedShape implements RectangularShape {
     public double getSize(){return size;}
 
     /** Set glyph's size by setting its bounding circle's radius.
-     *@see #reSize(float factor)
+     *@see #reSize(double factor)
      */
     public void sizeTo(double radius){/*XXX:TBW*/}
 
     /** Set glyph's size by multiplying its bounding circle radius by a factor. 
-     *@see #sizeTo(float radius)
+     *@see #sizeTo(double radius)
      */
     public void reSize(double factor){/*XXX:TBW*/}
